+++ apps/backend/src/modules/ai-assistant/libs/helpers/explain-task/generate-explain-task-prompt.ts
@@ -1,0 +1,22 @@
+import {
+	type OpenAIRequestMessage,
+	OpenAIRoleKey,
+} from "~/libs/modules/open-ai/open-ai.js";
+
+import { type TaskCreateDto } from "../../types/types.js";
+import { ExplainTaskPromptTemplate } from "./explain-task-prompt-template.js";
+
+function generateExplainTaskPrompt(task: TaskCreateDto): OpenAIRequestMessage {
+	const content = `
+	{
+	"context": "${ExplainTaskPromptTemplate.EXPLAIN_TASK_CONTEXT}",
+	"task": ${JSON.stringify(task)},
+	}`;
+
+	return {
+		content,
+		role: OpenAIRoleKey.USER,
+	};
+}
+
+export { generateExplainTaskPrompt };