+++ apps/mobile/src/libs/components/app/libs/constants/constants.ts
@@ -1,0 +1,7 @@
+import { config } from "~/libs/packages/config/config";
+
+const { HOST, SCHEME } = config.ENV.API;
+
+const DEEP_LINK_PREFIX = `${SCHEME}://${HOST}`;
+
+export { DEEP_LINK_PREFIX };