+++ apps/mobile/src/libs/components/typing-text-view/libs/enums/typing-text-animation.ts
@@ -1,0 +1,7 @@
+const TypingTextAnimation = {
+	DEFAULT_START_TYPING_DELAY: 0,
+	LETTERS_PRINTED_PER_STEP: 1,
+	NEXT_LETTERS_PRINTED_DELAY: 10,
+} as const;
+
+export { TypingTextAnimation };