+++ apps/mobile/src/libs/enums/app/base-color.enum.ts
@@ -9,10 +9,14 @@ const BaseColor = {
 	DARK_GRAY: "rgba(124, 124, 124, 1)",
 	EXTRA_LIGHT_GRAY: "rgba(224, 224, 224, 0.3)",
 	GRAY: "rgba(176, 176, 178, 1)",
-	GREEN: "rgba(194, 234, 205, 1)",
+	GREEN: "rgba(32, 184, 47, 1)",
 	LIGHT_GRAY: "rgba(224, 224, 224, 1)",
 	LIGHT_GREEN: "rgba(240, 255, 244, 1)",
 	RED: "rgba(255, 41, 40, 1)",
+	TRANSPARENT_BLACK: "rgba(0, 0, 0, 0.5)",
+	TRANSPARENT_BLUE: "rgba(64, 190, 255, 0.2)",
+	TRANSPARENT_GREEN: "rgba(32, 184, 47, 0.2)",
+	TRANSPARENT_RED: "rgba(255, 41, 40, 0.2)",
 } as const;
 
 export { BaseColor };