+++ apps/mobile/src/libs/types/navigation/questions-stack-navigation-parameter-list.type.ts
@@ -2,6 +2,7 @@ import { type QuestionsStackName } from "~/libs/enums/enums";
 
 type QuestionsStackNavigationParameterList = {
 	[QuestionsStackName.BOTTOM_TABS]: undefined;
+	[QuestionsStackName.INTRODUCTION]: undefined;
 	[QuestionsStackName.NOTIFICATION_QUESTIONS]: undefined;
 	[QuestionsStackName.ONBOARDING]: undefined;
 	[QuestionsStackName.QUIZ]: undefined;